mlflow-requirements.txt
# MLflow Server Requirements
# Minimal requirements for MLflow server on Google Cloud Run

flask>=2.3.0
google-auth>=2.40.3
google-cloud-storage>=2.10.0
gunicorn>=21.0.0
mlflow>=2.8.0
